+++ src/main/java/fpt/project/bsmart/service/IUserService.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 
 public interface IUserService {
-    Integer saveUser(CreateAccountRequest createAccountRequest);
+    Long saveUser(CreateAccountRequest createAccountRequest);
 
     Optional<User> findByUsername(String username);
 }
